+++ mockall_derive/src/mock_function.rs
@@ -1574,6 +1574,14 @@ impl<'a> ToTokens for Context<'a> {
         #[cfg(feature = "nightly_derive")]
         let must_use = quote!();
 
+        #[cfg(not(feature = "nightly_derive"))]
+        let clear_poison = quote!();
+        #[cfg(feature = "nightly_derive")]
+        let clear_poison = quote!(
+            #[allow(clippy::incompatible_msrv)]
+            get_expectations().clear_poison();
+        );
+
         quote!(
             /// Manages the context for expectations of static methods.
             ///
@@ -1621,6 +1629,9 @@ impl<'a> ToTokens for Context<'a> {
             impl #ty_ig Drop for Context #ty_tg #ty_wc {
                 fn drop(&mut self) {
                     if ::std::thread::panicking() {
+                        // Clear poison since we're about to clear the Mutex's
+                        // contents anyway.
+                        #clear_poison
                         // Drain all expectations so other tests can run with a
                         // blank slate.  But ignore errors so we don't
                         // double-panic.